backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic[email]==2.9.0
pydantic-settings==2.5.0
supabase==2.9.1
bcrypt==4.2.0
PyJWT==2.9.0
google-generativeai==0.8.0
google-auth==2.35.0
requests==2.32.0
python-multipart==0.0.9
python-dotenv==1.0.1
